+++ src/includes/classes/Utils/Content.php
@@ -32,7 +32,7 @@ class Content extends SCoreClasses\SCore\Base\Core
     /**
      * Tokenizers.
      *
-     * @since 16xxxx Content utils.
+     * @since 160722.57445 Content utils.
      *
      * @type array|null
      */
@@ -41,7 +41,7 @@ class Content extends SCoreClasses\SCore\Base\Core
     /**
      * Shortcode tag name.
      *
-     * @since 16xxxx Content utils.
+     * @since 160722.57445 Content utils.
      *
      * @param string
      */
@@ -50,7 +50,7 @@ class Content extends SCoreClasses\SCore\Base\Core
     /**
      * Class constructor.
      *
-     * @since 16xxxx Content utils.
+     * @since 160722.57445 Content utils.
      *
      * @param Classes\App $App Instance.
      */