+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## $v
+## v160722.57445
 
 - Bug fix. Improve filter-state backup handler.
 - Bug fix. Improve nested if block-level enforcement handler. Account for `^` in regex.